+++ R/logger-meta.R
@@ -15,6 +15,7 @@ logger_meta_env <- function(log_level = NULL,
   delayedAssign("fn", deparse_to_one_line(.topcall[[1]]), assign.env = env)
   delayedAssign("call", deparse_to_one_line(.topcall), assign.env = env)
   delayedAssign("topenv", top_env_name(.topenv), assign.env = env)
+  delayedAssign("location", log_call_location(.logcall), assign.env = env)
 
   env$time <- timestamp
   env$levelr <- log_level